channel.c: Fix conversion inconsistency
ch_log expects an int (using %d format specifier) but got a long
(linenr_T)

+++ src/channel.c
@@ -1443,7 +1443,7 @@ channel_write_in(channel_T *channel)
     }
     else
 	ch_log(channel, "Still %d more lines to write",
-					  buf->b_ml.ml_line_count - lnum + 1);
+				    (int)buf->b_ml.ml_line_count - lnum + 1);
 }
 
 /*
@@ -1537,7 +1537,7 @@ channel_write_new_lines(buf_T *buf)
 		ch_log(channel, "written %d lines to channel", written);
 	    if (lnum < buf->b_ml.ml_line_count)
 		ch_log(channel, "Still %d more lines to write",
-					      buf->b_ml.ml_line_count - lnum);
+					(int)buf->b_ml.ml_line_count - lnum);
 
 	    in_part->ch_buf_bot = lnum;
 	}
@@ -2081,7 +2081,7 @@ channel_get_json(
 	{
 	    *rettv = item->jq_value;
 	    if (tv->v_type == VAR_NUMBER)
-		ch_log(channel, "Getting JSON message %d", tv->vval.v_number);
+		ch_log(channel, "Getting JSON message %d", (int)tv->vval.v_number);
 	    remove_json_node(head, item);
 	    return OK;
 	}
